refactor: 기도 완료 로직 도메인 변경 #33
- notification 에서 prayers로 변경
