+++ src/domain/members/services/memberService.ts
@@ -1,6 +1,4 @@
 import apiService from "../../../common/apis/apiService";
-import type { MessageResponse } from "../../../common/types/messageResponse";
-import { NotifyPrayerRequest } from "../../notifications/types/request/notifyPrayerRequest";
 
 export const memberService = {
   // 사용자 프로필 단일 정보 조회
