--- src/domain/notifications/constants/notificationType.ts
@@ -1,3 +1,0 @@
-export enum NOTIFICATION_TYPE {
-  PRAYER_COMPLETION = "PRAYER_COMPLETION",
-}
