--- src/domain/notifications/types/request/notifyPrayerRequest.ts
@@ -1,7 +1,0 @@
-import { NOTIFICATION_TYPE } from "../../constants/notificationType";
-
-export interface NotifyPrayerRequest {
-  prayerTitleId: number | null;
-  roomId: number | null;
-  type: NOTIFICATION_TYPE;
-}
