+++ src/domain/prayers/types/request/createPrayerCompletionRequest.ts
@@ -1,0 +1,3 @@
+export interface CreatePrayerCompletionRequest {
+  roomId: number;
+}
